requirements.txt
flask
plotly
matplotlib
tensorflow
pillow
numpy
opencv-python
werkzeug